Add extra log output to track down leaked socket fd problem
file descriptor number is logged when an SMTP connection is accepted
and when it is about to be closed.

Also moved the error handler in do-smtp higher up so that it is still
in effect when the unwind-protect is operating (so that errors that
occur in the unwind-protect cleanup forms don't go unseed).

Also log accept-connection errors to see if they could possibly be
related to the leak.