v1.2.29: make 9.0 SMP the default build lisp
Testing of the SMP shows that it works, so we tentatively put it into
service during an email blast and it's holding up really well.  No
problems noticed and it solved one of the biggest problems of the
blasts on freon (the machine that delivers all the email from
meatlocker): non-blast emails get through without error.